+++ app/components/magicui/aurora-text.tsx
