+++ app/login/page.tsx
@@ -1,6 +1,6 @@
 import { LoginForm } from "@/components/ui/login-form"
 import { cn } from "@/lib/utils";
-import { DotPattern } from "../../components/magicui/dot-pattern";
+import { DotPattern } from "../components/magicui/dot-pattern";
 
 export default function Page() {
   return (
